wip refactoring, TODO: update parent components on change, add rolling recording to refactoring
