+++ src/GeneralSettings.tsx
@@ -1,0 +1,106 @@
+import {
+	ButtonItem,
+	PanelSectionRow,
+	ServerAPI,
+	Dropdown,
+	DropdownOption,
+	SingleDropdownOption
+} from "decky-frontend-lib";
+
+import {
+	VFC,
+	useState,
+	useEffect
+} from "react";
+
+
+export const GeneralSettings: VFC<{ serverAPI: ServerAPI}> = ({serverAPI}) => {
+
+	const [isCapturing, setCapturing] = useState<boolean>(false);
+	const [isRolling, setRolling] = useState<boolean>(false);
+
+	const [localFilePath, setLocalFilePath] = useState<string>("/home/deck/Videos");
+
+	const formatOptionMp4 = { data: "mp4", label: "MP4" } as SingleDropdownOption
+	const formatOptionMkv = { data: "mkv", label: "Matroska (.mkv)" } as SingleDropdownOption;
+	const formatOptionMov = { data: "mov", label: "QuickTime (.mov)" } as SingleDropdownOption;
+	const formatOptions: DropdownOption[] = [formatOptionMkv, formatOptionMp4, formatOptionMov];
+	const [localFileFormat, setLocalFileFormat] = useState<DropdownOption>(formatOptionMp4);
+
+	const initState = async () => {
+
+		const getIsCapturingResponse = await serverAPI.callPluginMethod('is_capturing', {});
+		setCapturing(getIsCapturingResponse.result as boolean);
+
+		const getIsRollingResponse = await serverAPI.callPluginMethod('is_rolling', {});
+		setRolling(getIsRollingResponse.result as boolean);
+
+		const getLocalFilepathResponse = await serverAPI.callPluginMethod('get_local_filepath', {})
+		setLocalFilePath(getLocalFilepathResponse.result as string);
+
+		const getLocalFileFormatResponse = await serverAPI.callPluginMethod('get_local_fileformat', {})
+		const localFileFormatResponseString: string = getLocalFileFormatResponse.result as string;
+		if (localFileFormatResponseString == "mp4") {
+			setLocalFileFormat(formatOptionMp4)
+		} else if (localFileFormatResponseString == "mkv") {
+			setLocalFileFormat(formatOptionMkv)
+		} else if (localFileFormatResponseString == "mov") {
+			setLocalFileFormat(formatOptionMov)
+		} else {
+			// should never happen? default back to mp4
+			setLocalFileFormat(formatOptionMp4)
+		}
+
+	}
+
+	const pickFolder = async () => {
+		const filePickerResponse = await serverAPI.openFilePicker(localFilePath, false);
+		setLocalFilePath(filePickerResponse.path)
+		await serverAPI.callPluginMethod('set_local_filepath', {localFilePath: filePickerResponse.path});
+	}
+
+	const disableFileformatDropdown = () => {
+		if (isCapturing) {
+			return true;
+		}
+		if (isRolling) {
+			return true;
+		}
+		return false;
+	}
+
+	useEffect(() => {
+		initState();
+	}, []);
+
+	return (
+		<PanelSectionRow>
+
+			<div>Settings</div>
+			<ButtonItem
+				label="Path"
+				disabled={disableFileformatDropdown()}
+				bottomSeparator="none"
+				layout="below"
+				onClick={() => {
+					pickFolder();
+				}}>
+				{localFilePath}
+			</ButtonItem>
+
+			<Dropdown
+				menuLabel="Select the video file format"
+				disabled={disableFileformatDropdown()}
+				strDefaultLabel={localFileFormat.label as string}
+				rgOptions={formatOptions}
+				selectedOption={localFileFormat}
+				onChange={(newLocalFileFormat) => {
+					serverAPI.callPluginMethod('set_local_fileformat', { fileformat: newLocalFileFormat.data });
+					setLocalFileFormat(newLocalFileFormat);
+				}}
+			/>
+
+		</PanelSectionRow>
+	);
+
+};
